feat: restore wallpapers repo
Closes #39 by restoring the old wallpapers.
Like discussed in that issue, the old git history won't be restored, but
original authors are credited in the READMEs.

Wallpapers which weren't OC have been removed from the repository, some
still link to the artist's website if we were able to find the link.

From now on, we will require stricter checks for OC content.

Co-authored-by: ghostx31 <himanshu312001@gmail.com>